lib: fix reinstall when Depends field has different format
This fixed issues that package is reinstalled if Depends field has a
different formating (spaces) than what is specified in the package. This
is caused by OPKG formating it automatically.
The solution here is to parse the dependency and compare parsed tree
instead. It is not exactly ideal solution as we have to parse
dependencies possibly multiple times but it is for now the easiest. The
ideal way would be to have here already parsed form we could compare
directly.