docs/changelog: fix invalid keyword for transaction end
There was a mismatch between documentation and the implementation where
documentation stated `STOP` as the keyword for transaction end and the
code instead was using `END`. In the end it doesn't matter which we
choose and it is easier for use to just tweak the documentation instead
changing the code now and thus from now on it is `END` and not `STOP`.