pkgupdate/arguments: Do not fail on obsolete options
Do not fail when `--state-log` or `--task-log` is provided. Output a
warning and ignore them instead for a sake of backward compatibility. We
encountered that in some conditions, updater might be run the old way
during migration from TOS 5.0 to TOS 6.0 and this should prevent
errors and half-installations in such cases.