Include a way how to get a serial number
This is a dependency for staged updates. We need to somehow group
various routers together and gradually add more and more routers to
every new release. To do that we need a way how to uniquely identify
them and serial number is here precisely for that. This way we will be
able to use it in lua scripts and point right batch of routers to the
right direction.